탈퇴 시 해당 계정의 게시글 삭제

+++ src/main/java/com/example/prjjsp20241022/service/MemberService.java
@@ -1,6 +1,7 @@
 package com.example.prjjsp20241022.service;
 
 import com.example.prjjsp20241022.dto.Member;
+import com.example.prjjsp20241022.mapper.BoardMapper;
 import com.example.prjjsp20241022.mapper.MemberMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,6 +14,7 @@ import java.util.List;
 @Transactional
 public class MemberService {
     private final MemberMapper mapper;
+    private final BoardMapper boardMapper;
 
     public void addMember(Member member) {
         mapper.insert(member);
@@ -26,9 +28,17 @@ public class MemberService {
         return mapper.selectById(id);
     }
 
-    public boolean removeMember(String id, String password) {
-        int count = mapper.deleteByIdAndPsd(id, password);
-        return count == 1;
+    public boolean remove(String id, String password) {
+//        int count = mapper.deleteByIdAndPsd(id, password);
+        //게시물 board 먼저 삭제 후
+        //회원 member 삭제
+        int cnt = 0;
+        Member member = mapper.selectById(id);
+        if (member.getPassword().equals(password)) {
+            boardMapper.deleteByMemberId(id);
+            cnt = mapper.deleteByIdAndPsd(id, password);
+        }
+        return cnt == 1;
     }
 
     public void update(Member member) {
